+++ MK4duo/Configuration_Temperature.h
@@ -39,6 +39,7 @@
  * - Inverted PINS
  * - Thermal runaway protection
  * - Prevent cold extrusion
+ * - Safety timer
  *
  */
 
@@ -429,6 +430,7 @@
 /***********************************************************************
  ************************ Prevent cold extrusion ***********************
  ***********************************************************************
+ *                                                                     *
  * This option prevents extrusion if the temperature is                *
  * below EXTRUDE MINTEMP.                                              *
  * Add M302 to set the minimum extrusion temperature and/or turn       *
@@ -444,4 +446,22 @@
 // Note: For Bowden Extruders make this large enough to allow load/unload.
 //#define PREVENT_LENGTHY_EXTRUDE
 #define EXTRUDE_MAXLENGTH 400
+/***********************************************************************/
+
+
+/***********************************************************************
+ **************************** Safety Timer *****************************
+ ***********************************************************************
+ *                                                                     *
+ * Set safety timer expiration time in minutes, when not printing and  *
+ * safety timer expires all heater are set to zero.                    *
+ * Remeber put M530 S1 in your start gcode for start printing and      *
+ * M530 S0 in end gcode for stop printing.                             *
+ * M86 M[min] set safety timer expiration time                         *
+ * M86 M0 will disable safety timer.                                   *
+ *                                                                     *
+ *      IT IS HIGHLY RECOMMENDED TO LEAVE THIS OPTION ENABLED!         *
+ *                                                                     *
+ ***********************************************************************/
+#define SAFETYTIMER_TIME_MINS 30
 /***********************************************************************/