+++ Documentation/changelog.md
@@ -8,6 +8,8 @@
 * Add support thermocouples for bed and hotend.
 * Add menu axis limit to menu advanced.
 * Add Double-Quad Stepping to command M569 Q and save it into EEPROM.
+* Add Option for Safety Timer in configuration_temperature.h
+* Add M86 M[min] set safety timer expiration time in minute. M86 M0 will disable safety timer
 * M301 - Set PID parameters P I D and C.
 *   H[heaters] 0-5 Hotend, -1 BED, -2 CHAMBER, -3 COOLER
 *   T[int] 0-3 For Select Beds or Chambers (default 0)